chore(wing): use tsup instead of tsc
Improves the Console DX by allowing to work with the `@wingconsole/server` package without starting an additional dev script.
