+++ apps/wing-console/console/app/CONTRIBUTING.md
@@ -5,5 +5,3 @@
 ```sh
 pnpm dev
 ```
-
-Also, you have to start the dev script on the `apps/wing-console/console/server` package.
