+++ apps/wing/tsup.config.ts
@@ -1,12 +1,8 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
-  external: ["fsevents"],
+  entry: ["src/cli.ts"],
   format: ["cjs"],
   dts: true,
   clean: true,
-  env: {
-    LICENSE_FILE: "../LICENSE.md",
-  },
 });
